drivers/system76/dgpu: Add driver for System76 switchable graphics
This adds a driver for System76 systems with switchable NVIDIA
graphics. The driver provides ACPI support for dynamically
powering on and off the GPU, a function for enabling the GPU
power in the bootblock, and a hook in ramstage for moving GPU
prefetch resources above 4 GiB.

Tested on system76/addw2, system76/gaze15, and system76/oryp6,
which will be submitted in future patches.

Signed-off-by: Jeremy Soller <jeremy@system76.com>
Change-Id: I05833e88457bfc3612cc9a8cbf1eec68dcfc5566